Trigger win64 extension build before win32 build of Slicer
From Andras:

As I see on the dashboard today, the problem is that the win32
application is built before the win64 extensions (win64 application,
win32 application, win64 extensions, win32 extensions). Effectively,
the win32 application build makes the win64 extensions become available
 at 1pm instead of 7am.

Could you please change the build order to: win64 application, win64
extensions, win32 application, win32 extensions?
It would decrease the time window when an application package can be
downloaded already but extensions are not yet available (for both win64
and win32).

Discussion: http://slicer-devel.65872.n3.nabble.com/Extensions-are-not-available-in-the-morning-tt4029422.html