factory-south-win7 - Re-enable building of CLIs
While working on Slicer commit r22241 [1], the dashboard scripts where
locally modified so that only the packaging step was done. Considering
the script remained locally modified, the nightly build for the following
two nights were not properly done.

For some reason to be explained, manually starting the continuous
build (after restoring the script) prevented the result from being
reported on the dashboard.

This is why the CLI have been re-enabled by this commit for the
windows7 experimental build.

[1] http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=22241