Build nightly extensions before 4.5 ones
As explained by @jcfr, with the current setup, all 4.5 extensions are built every
night, even when there are no updates to 4.5 extensions index. This prevents the
nightly extension, which are more likely to change, from being built at all.

Lacking automatic detection of the 4.5 extensions index changes and building of
the modified extensions only, it is proposed that the order between 4.5 and nightly
is changed manually on as needed basis. It is expected that nightly builds will be
changing at a faster pace than 4.5, thus it is natural to have nightly built in the
first place.