Revert "Build 4.5 extensions before the nightly ones."
This reverts commit 1fbf4c4e337e2242b3e1d266edd6ec99b699e163.

With the current setup, all 4.5 extensions are built every night, even
when there are no updates to 4.5 extensions index. This prevents the
nightly extension, which are more likely to change, from being built
at all.

Lacking automatic detection of the 4.5 extensions index changes and
building of the modified extensions only, it is proposed that the order
between 4.5 and nightly is changed manually on as needed basis. It is
expected that nightly builds will be changing at a faster pace than 4.5,
thus it is natural to have nightly built in the first place.

In a follow-up commit, the build script will be updated so that 4.5
are build before the nightly ones only during the weekend.

Co-authored-by: Andriy Fedorov <fedorov@bwh.harvard.edu>