style: Consistently reference the same macOS deployment target
This changes the macOS deployment target to "14.0" across various configuration files. While the variable is not used on Windows and Linux dashboards, this will avoid confusion, reduce the chance of copy-paste issues, and streamline file comparison.

computron-slicerextensions_stable_nightly.cmake is maintained as 13.0 because Slicer 5.10.0 stable was built targeting 13.0. computron-slicer_stable_package.cmake is maintain at 13.0 for consistency as well.